issue #1 반영
